Add optional dfdt parameter to dedrift methods
The drift rate value passed to `dedrift` is divided by the new `dfdt`
parameter to calculate the normalized drift rate used by the underlying
routines.  If `dfdt` is not given it defaults to 1, which implies that
the given drift rate value is already normalized.  If the give drift
rate is in Hz/s, then `dfdt` should be the frequency step size (in Hz)
divided by the time step size (in seconds).